Move indexing code to a new file
